tools/xl: Add xl fork-vm command
Adding the xl fork-vm command, compiled only on x86. Only the essential bits
are available via this command to create a fork and launch QEMU for it. The
command still allows to perform the task in a split-model, first creating the
fork and launching QEMU only later.

Signed-off-by: Tamas K Lengyel <tamas.lengyel@intel.com>
Message-Id: <da122402f491598df80517a458e050ec46f640ac.1591017086.git.tamas.lengyel@intel.com>
Reviewed-by he sent last week