docs/support-matrix: Gross bodge to unbreak docs rendering
The cronjob which renders https://xenbits.xen.org/docs/ has been broken for a
while.  commitish_version() pulls an old version of xen/Makefile out of
history, and uses the xenversion rule.

Currently, this fails with:

  tmp.support-matrix.xen.make:130: scripts/Kbuild.include: No such file or directory

which is because the Makefile legitimately references Kbuild.include with a
relative rather than absolute path.

Rearrange $CWD of the make rune to be in xen/

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20200604205226.14518-1-andrew.cooper3@citrix.com>