xen/x86: irq: Avoid a TOCTOU race in pirq_spin_lock_irq_desc()
Even if we assigned pirq->arch.irq to a variable, a compile is still
allowed to read pirq->arch.irq multiple time. This means that the value
checked may be different from the value used to get the desc.

Force the compiler to only do one read access by using read_atomic().

Signed-off-by: Julien Grall <jgrall@amazon.com>
Message-Id: <20200722165300.22655-1-julien@xen.org>